backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.0
pytest==8.4.1
pytest-django==4.11.1
